+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Please ADD ALL Changes to the UNRELASED SECTION and not a specific release
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.16.0.25740
 - FF-1429 - Updated FunFair.Test.Common to 5.0.0.735
 - FF-1429 - Updated AsyncFixer to 1.4.0
+- FF-1429 - Updated FunFair.Test.Common to 5.1.0.784
 ### Removed
 ### Deployment Changes