+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Fixed
 ### Changed
 - FF-1429 - Updated SonarAnalyzer.CSharp to 8.23.0.32424
+- FF-1429 - Updated Microsoft.NET.Test.Sdk to 16.10.0
 ### Removed
 ### Deployment Changes